提交

+++ src/com/yq/dao/OrderEvalDao.java
@@ -9,15 +9,15 @@ public interface OrderEvalDao {
 
 	Long insertEval(OrderEval orderEval);
 	
-	List<OrderEval> getAllEvalByGoodId(Long goodId);
+	List<OrderEval> getAllEvalByGoodId(Integer goodId);
 	
 	List<OrderEval> getAllEvalByGoodIdOpenId(Map<String, Object> map);
 	
-	int getCountByGoodId(Long goodId);
+	int getCountByGoodId(Integer goodId);
 	
 	//查找好评的个数
-	int getGoodCountByGoodId(Long goodId);
+	int getGoodCountByGoodId(Integer goodId);
 	
 	//查找差评的个数
-	int getBadCountByGoodId(Long goodId);
+	int getBadCountByGoodId(Integer goodId);
 }
